.history/requirements_20220520002634.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
click==8.1.3
flask==2.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
zipp==3.8.0
WTForms-3.0.1 flask-wtf-1.0.1